requirements.txt
langchain==0.1.11
openai
google-cloud-translate==2.0.1
python-dotenv==1.0.1
pandas
requests
